fix: minor improvements, restructure remaining files
Implemented minor enhancements, relocated remaining files not aligning with the new project structure. Refactored copy-pasted comments and removed unused components.

Asana: https://app.asana.com/0/1203851531040615/1206687927869186/f
