+++ src/Asv.Avalonia.ToolkitGallery/Asv.Avalonia.ToolkitGallery.csproj
@@ -68,5 +68,9 @@
       <LastGenOutput>RS.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Controls\" />
+  </ItemGroup>
   
 </Project>
