--- src/Asv.Avalonia.ToolkitGallery/Controls/LinksCardItem.axaml
@@ -1,50 +1,0 @@
-<Styles xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:controls="clr-namespace:Asv.Avalonia.ToolkitGallery.Controls">
-    <Styles.Resources>
-        <Geometry x:Key="GithubIcon">M41.0098 0.61026C18.4186 0.61026 0.120728 19.0646 0.120728 41.8489C0.120728 60.097 11.8252 75.51 28.0787 80.9741C30.1231 81.335 30.8898 80.0978 30.8898 79.0153C30.8898 78.0359 30.8387 74.7883 30.8387 71.3346C20.5653 73.2419 17.9075 68.8087 17.0897 66.489C16.6297 65.3034 14.6364 61.6435 12.8986 60.6641C11.4675 59.8908 9.423 57.9836 12.8475 57.932C16.0675 57.8805 18.3675 60.9218 19.1342 62.159C22.8142 68.3963 28.692 66.6437 31.0431 65.5612C31.4009 62.8807 32.4742 61.0765 33.6498 60.0455C24.552 59.0145 15.0453 55.4577 15.0453 39.6839C15.0453 35.1992 16.6297 31.4877 19.2364 28.601C18.8275 27.57 17.3964 23.3431 19.6453 17.6728C19.6453 17.6728 23.0697 16.5902 30.8898 21.8997C34.1609 20.9719 37.6365 20.5079 41.1121 20.5079C44.5877 20.5079 48.0632 20.9719 51.3344 21.8997C59.1544 16.5387 62.5789 17.6728 62.5789 17.6728C64.8278 23.3431 63.3966 27.57 62.9878 28.601C65.5944 31.4877 67.1789 35.1476 67.1789 39.6839C67.1789 55.5092 57.6211 59.0145 48.5232 60.0455C50.0055 61.3342 51.2832 63.8085 51.2832 67.6747C51.2832 73.1903 51.2321 77.6235 51.2321 79.0153C51.2321 80.0978 51.9988 81.3865 54.0433 80.9741C62.1609 78.2111 69.2149 72.9498 74.2119 65.9311C79.209 58.9123 81.8976 50.4897 81.899 41.8489C81.899 19.0646 63.6011 0.61026 41.0098 0.61026Z</Geometry>
-    </Styles.Resources>
-    
-    <Design.PreviewWith>
-        <Border Padding="20" Height="200">
-            <controls:LinksCardItem></controls:LinksCardItem>
-        </Border>
-       </Design.PreviewWith>
-    
-    <Style Selector="controls|LinksCardItem">
-        <Setter Property="Background" Value="Transparent"/>
-        <Setter Property="CornerRadius" Value="30"/>
-        <Setter Property="Template">
-            <ControlTemplate>
-                <Viewbox>
-                    <Button>
-                         <Border Height="150" 
-                                 Width="200"
-                                 CornerRadius="30"
-                                 Background="#D9D9D9">
-                              <Canvas>
-                                  <PathIcon Canvas.Top="30"
-                                  Canvas.Left="50" 
-                                  Width="100" 
-                                  Height="50" 
-                                  Foreground="Black" 
-                                  Data="{StaticResource GithubIcon}"/>
-                                  <Border Canvas.Top="100"
-                                          Height="50"
-                                          Width="200"
-                                          CornerRadius="0,0,30,30" Background="#4F5B66"
-                                          VerticalAlignment="Bottom"
-                                          HorizontalAlignment="Center">
-                                      <TextBlock VerticalAlignment="Center"
-                                                 HorizontalAlignment="Center"
-                                                 Text="Project GitHub"/>  
-                                  </Border>
-                              </Canvas>
-                         </Border>
-                    </Button>
-                </Viewbox>
-            </ControlTemplate>
-        </Setter>
-       
-    </Style>
-</Styles>
