--- src/Asv.Avalonia.ToolkitGallery/Controls/LinksCardItem.axaml.cs
@@ -1,11 +1,0 @@
-using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Controls.Primitives;
-using Avalonia.Markup.Xaml;
-
-namespace Asv.Avalonia.ToolkitGallery.Controls;
-
-public class LinksCardItem : TemplatedControl
-{
-   
-}
