+++ src/Asv.Avalonia.ToolkitGallery/Services/ServiceWithConfigBase.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Asv.Avalonia.ToolkitGallery.Models;
+using Asv.Avalonia.ToolkitGallery.Tools;
 using Asv.Cfg;
 
 namespace Asv.Avalonia.ToolkitGallery.Services;
