+++ src/Asv.Avalonia.ToolkitGallery/ViewModels/Pages/SettingsViewModel.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reactive.Linq;
-using Asv.Avalonia.ToolkitGallery.Models;
 using Asv.Avalonia.ToolkitGallery.Services.Theme;
+using Asv.Avalonia.ToolkitGallery.Tools;
 using Asv.Common;
 using Microsoft.Extensions.DependencyInjection;
 using ReactiveUI;
